--- a/nntrainer/layers/bn_layer.cpp
+++ b/nntrainer/layers/bn_layer.cpp
@@ -423,7 +423,8 @@ void BatchNormalizationLayer::read(std::ifstream &file,
                                    ml::train::ExecutionMode mode,
                                    bool trainable,
                                    TensorDim::DataType definedWeightDataType,
-                                   bool fsu, size_t start_offset) {
+                                   bool fsu, size_t start_offset,
+                                   bool read_from_offset) {
   if (opt_var) {
     for (unsigned int i = 0; i < run_context.getNumWeights(); ++i) {
       if (run_context.isGradientLastAccess(i) && trainable) {