--- a/nntrainer/layers/bn_layer.h
+++ b/nntrainer/layers/bn_layer.h
@@ -144,7 +144,7 @@ public:
   void read(std::ifstream &file, RunLayerContext &context, bool opt_var,
             ml::train::ExecutionMode mode, bool trainable,
             TensorDim::DataType definedWeightDataType, bool fsu = false,
-            size_t start_offset = -1, bool read_from_offset = false) override;
+            size_t start_offset = 0, bool read_from_offset = false) override;
 
 private:
   float divider; /**< size of the axes of the reduced */