--- a/nntrainer/layers/layer_node.cpp
+++ b/nntrainer/layers/layer_node.cpp
@@ -480,12 +480,12 @@ void LayerNode::exportTo(Exporter &exporter,
 }
 
 void LayerNode::read(std::ifstream &file, bool opt_var,
-                     ml::train::ExecutionMode mode, bool fsu) {
+                     ml::train::ExecutionMode mode, bool fsu, size_t start_offset) {
   NNTR_THROW_IF(!run_context, std::runtime_error)
     << __func__ << " layer needs to be finalized first!";
   getLayer()->read(file, *run_context, opt_var, mode,
                    (getTrainable() && mode == ml::train::ExecutionMode::TRAIN),
-                   getWeightDataType(), fsu);
+                   getWeightDataType(), fsu, start_offset);
 }
 
 void LayerNode::save(std::ofstream &file, bool opt_var,