--- a/nntrainer/layers/layer_node.cpp
+++ b/nntrainer/layers/layer_node.cpp
@@ -481,12 +481,12 @@ void LayerNode::exportTo(Exporter &exporter,
 
 void LayerNode::read(std::ifstream &file, bool opt_var,
                      ml::train::ExecutionMode mode, bool fsu,
-                     size_t start_offset) {
+                     size_t start_offset, bool read_from_offset) {
   NNTR_THROW_IF(!run_context, std::runtime_error)
     << __func__ << " layer needs to be finalized first!";
   getLayer()->read(file, *run_context, opt_var, mode,
                    (getTrainable() && mode == ml::train::ExecutionMode::TRAIN),
-                   getWeightDataType(), fsu, start_offset);
+                   getWeightDataType(), fsu, start_offset, read_from_offset);
 }
 
 void LayerNode::save(std::ofstream &file, bool opt_var,