--- a/nntrainer/tensor/bcq_tensor.h
+++ b/nntrainer/tensor/bcq_tensor.h
@@ -216,7 +216,8 @@ public:
   /**
    * @copydoc Tensor::read(std::ifstream &file)
    */
-  void read(std::ifstream &file) override;
+  void read(std::ifstream &file, size_t start_offset,
+            bool read_from_offset) override;
 
   /**
    * @copydoc Tensor::argmax()
@@ -231,7 +232,8 @@ public:
   /**
    * @copydoc TensorBase::read_quantization_info(std::ifstream &file)
    */
-  void read_quantization_info(std::ifstream &file) override;
+  void read_quantization_info(std::ifstream &file, size_t start_offset,
+                              bool read_from_offset) override;
 
   /**
    * @copydoc TensorBase::size()