--- a/nntrainer/tensor/char_tensor.cpp
+++ b/nntrainer/tensor/char_tensor.cpp
@@ -420,7 +420,7 @@ void CharTensor::save(std::ostream &file) {
   putData();
 }
 
-void CharTensor::read(std::ifstream &file) {
+void CharTensor::read(std::ifstream &file, size_t start_offset) {
   /// @note Read quantization information
   read_quantization_info(file);