--- a/nntrainer/tensor/char_tensor.cpp
+++ b/nntrainer/tensor/char_tensor.cpp
@@ -423,7 +423,7 @@ void CharTensor::save(std::ostream &file) {
 void CharTensor::read(std::ifstream &file, size_t start_offset,
                       bool read_from_offset) {
   /// @note Read quantization information
-  read_quantization_info(file);
+  read_quantization_info(file, start_offset, read_from_offset);
 
   std::streamsize sz = static_cast<std::streamsize>(getMemoryBytes());
 
@@ -593,7 +593,7 @@ void CharTensor::save_quantization_info(std::ostream &file) {
                "[CharTensor::save] failed to write quantization information");
 }
 
-void CharTensor::read_quantization_info(std::ifstream &file) {
+void CharTensor::read_quantization_info(std::ifstream &file, size_t start_offset, bool read_from_offset) {
   checkedRead(file, (char *)&qscheme, sizeof(uint16_t),
               "[CharTensor::read] failed to read quantization information");
 }