--- a/nntrainer/tensor/char_tensor.cpp
+++ b/nntrainer/tensor/char_tensor.cpp
@@ -593,7 +593,9 @@ void CharTensor::save_quantization_info(std::ostream &file) {
                "[CharTensor::save] failed to write quantization information");
 }
 
-void CharTensor::read_quantization_info(std::ifstream &file, size_t start_offset, bool read_from_offset) {
+void CharTensor::read_quantization_info(std::ifstream &file,
+                                        size_t start_offset,
+                                        bool read_from_offset) {
   checkedRead(file, (char *)&qscheme, sizeof(uint16_t),
               "[CharTensor::read] failed to read quantization information");
 }