--- a/nntrainer/tensor/char_tensor.cpp
+++ b/nntrainer/tensor/char_tensor.cpp
@@ -431,8 +431,13 @@ void CharTensor::read(std::ifstream &file, size_t start_offset,
     << "read size: " << getMemoryBytes()
     << " is too big. It cannot be represented by std::streamsize";
 
+  if (read_from_offset) {
+    start_offset += sizeof(uint16_t);
+  }
+
   checkedRead(file, (char *)getData(), sz,
-              "[CharTensor::read] operation failed");
+              "[CharTensor::read] operation failed", start_offset,
+              read_from_offset);
   putData();
 }
 
@@ -597,7 +602,8 @@ void CharTensor::read_quantization_info(std::ifstream &file,
                                         size_t start_offset,
                                         bool read_from_offset) {
   checkedRead(file, (char *)&qscheme, sizeof(uint16_t),
-              "[CharTensor::read] failed to read quantization information");
+              "[CharTensor::read] failed to read quantization information",
+              start_offset, read_from_offset);
 }
 
 } // namespace nntrainer