--- a/nntrainer/tensor/char_tensor.h
+++ b/nntrainer/tensor/char_tensor.h
@@ -290,7 +290,8 @@ public:
   /**
    * @copydoc TensorBase::read_quantization_info()
    */
-  void read_quantization_info(std::ifstream &file) override;
+  void read_quantization_info(std::ifstream &file, size_t start_offset,
+                        bool read_from_offset ) override;
 
   /**
    * @copydoc Tensor::scale_size()