--- a/nntrainer/tensor/char_tensor.h
+++ b/nntrainer/tensor/char_tensor.h
@@ -244,7 +244,8 @@ public:
   /**
    * @copydoc Tensor::read(std::ifstream &file)
    */
-  void read(std::ifstream &file, size_t start_offset) override;
+  void read(std::ifstream &file, size_t start_offset,
+            bool read_from_offset) override;
 
   /**
    * @copydoc Tensor::argmax()