--- a/nntrainer/tensor/int4_tensor.cpp
+++ b/nntrainer/tensor/int4_tensor.cpp
@@ -345,7 +345,7 @@ void Int4QTensor::save(std::ostream &file) {
   putData();
 }
 
-void Int4QTensor::read(std::ifstream &file) {
+void Int4QTensor::read(std::ifstream &file, size_t start_offset) {
   /// @note Read quantization information
   read_quantization_info(file);