--- a/nntrainer/tensor/int4_tensor.cpp
+++ b/nntrainer/tensor/int4_tensor.cpp
@@ -345,7 +345,8 @@ void Int4QTensor::save(std::ostream &file) {
   putData();
 }
 
-void Int4QTensor::read(std::ifstream &file, size_t start_offset) {
+void Int4QTensor::read(std::ifstream &file, size_t start_offset,
+                       bool read_from_offset) {
   /// @note Read quantization information
   read_quantization_info(file);