--- a/nntrainer/tensor/int4_tensor.cpp
+++ b/nntrainer/tensor/int4_tensor.cpp
@@ -555,8 +555,9 @@ void Int4QTensor::save_quantization_info(std::ostream &file) {
                "[Int4QTensor::save] failed to write quantization information");
 }
 
-void Int4QTensor::read_quantization_info(std::ifstream &file, size_t start_offset,
-                        bool read_from_offset) {
+void Int4QTensor::read_quantization_info(std::ifstream &file,
+                                         size_t start_offset,
+                                         bool read_from_offset) {
   checkedRead(file, (char *)&qscheme, sizeof(uint16_t),
               "[Int4QTensor::read] failed to read quantization information");
 }