--- a/nntrainer/tensor/int4_tensor.cpp
+++ b/nntrainer/tensor/int4_tensor.cpp
@@ -356,8 +356,13 @@ void Int4QTensor::read(std::ifstream &file, size_t start_offset,
     << "read size: " << getMemoryBytes()
     << " is too big. It cannot be represented by std::streamsize";
 
+  if (read_from_offset) {
+    start_offset += sizeof(uint16_t);
+  }
+
   checkedRead(file, (char *)getData(), sz,
-              "[Int4QTensor::read] operation failed");
+              "[Int4QTensor::read] operation failed", start_offset,
+              read_from_offset);
   putData();
 }
 
@@ -559,7 +564,8 @@ void Int4QTensor::read_quantization_info(std::ifstream &file,
                                          size_t start_offset,
                                          bool read_from_offset) {
   checkedRead(file, (char *)&qscheme, sizeof(uint16_t),
-              "[Int4QTensor::read] failed to read quantization information");
+              "[Int4QTensor::read] failed to read quantization information",
+              start_offset, read_from_offset);
 }
 
 } // namespace nntrainer