--- a/nntrainer/tensor/int4_tensor.h
+++ b/nntrainer/tensor/int4_tensor.h
@@ -264,7 +264,8 @@ public:
   /**
    * @copydoc TensorBase::read_quantization_info()
    */
-  void read_quantization_info(std::ifstream &file) override;
+  void read_quantization_info(std::ifstream &file, size_t start_offset,
+                        bool read_from_offset) override;
 
   /**
    * @copydoc Tensor::getMemoryBytes()