--- a/nntrainer/tensor/int4_tensor.h
+++ b/nntrainer/tensor/int4_tensor.h
@@ -223,7 +223,8 @@ public:
   /**
    * @copydoc Tensor::read(std::ifstream &file)
    */
-  void read(std::ifstream &file, size_t start_offset) override;
+  void read(std::ifstream &file, size_t start_offset,
+            bool read_from_offset) override;
 
   /**
    * @copydoc Tensor::argmax()